+++ app/js/player/directives/track.js
@@ -6,7 +6,11 @@
  * @author SOON_
  */
 angular.module("FM.player.trackDirective", [
-
+    "FM.api.PlayerQueueResource",
+    "ui.bootstrap.popover",
+    "template/popover/popover-template.html",
+    "template/popover/popover.html",
+    "ui.bootstrap.dropdown"
 ])
 /**
  * @example
@@ -15,7 +19,8 @@ angular.module("FM.player.trackDirective", [
  * @class fmTrack
  */
 .directive("fmTrack",[
-    function (){
+    "PlayerQueueResource",
+    function (PlayerQueueResource){
         return {
             restrict: "EA",
             scope: {
@@ -24,7 +29,39 @@ angular.module("FM.player.trackDirective", [
                 user: "=",
                 timer: "=?"
             },
-            templateUrl: "partials/track.html"
+            templateUrl: "partials/track.html",
+            link: function($scope){
+
+                /**
+                 * @method onTrackUpdated
+                 * @param  {Object} track
+                 */
+                $scope.onTrackUpdate = function onTrackUpdated(track){
+                    if (track) {
+                        $scope.track.allArtists = "";
+                    }
+
+                    if (track && track.artists){
+                        angular.forEach(track.artists, function (artist, $index){
+                            var s = ($index !== track.artists.length - 1) ? (artist.name + ", ") : artist.name;
+                            $scope.track.allArtists += s;
+                        });
+                    }
+                };
+
+                /**
+                 * @method addToPlaylist
+                 * @param {Object} track
+                 */
+                $scope.addToPlaylist = function addToPlaylist (track) {
+                    if (track && track.uri) {
+                        PlayerQueueResource.save({ uri: track.uri });
+                    }
+                };
+
+                $scope.$watch("track", $scope.onTrackUpdate, true);
+
+            }
         };
     }
 ]);