+++ tests/e2e/protractor.conf.js
@@ -62,22 +62,22 @@ exports.config = {
   // https://code.google.com/p/selenium/wiki/DesiredCapabilities
   // and
   // https://code.google.com/p/selenium/source/browse/javascript/webdriver/capabilities.js
-  // capabilities: {
-  //   'browserName': 'chrome'
-  // },
+  capabilities: {
+    'browserName': 'firefox'
+  },
   //
   // install safari driver here:
   // https://code.google.com/p/selenium/issues/detail?id=7933#c33
 
-  multiCapabilities: [{
-    'browserName': 'firefox'
-  }, {
-    'browserName': 'chrome'
+  // multiCapabilities: [{
+  //   'browserName': 'firefox'
+  // }, {
+  //   'browserName': 'chrome'
   // }, {
   //   'browserName': 'safari'
   // }, {
   //   'browserName': 'internet explorer'
-  }],
+  // }],
 
   // ----- More information for your tests ----
   //