+++ tests/unit/playlist/controllers/PlaylistCtrl.js
@@ -3,7 +3,7 @@
 describe("FM.playlist.PlaylistCtrl", function() {
 
     var $rootScope, $location, $route, $scope, $q, $httpBackend, $notification, notificationSpy,
-        TracksResource, UsersResource, PlayerQueueResource, playlistData;
+        TracksResource, UsersResource, PlayerQueueResource, playlistData, playlistMeta;
 
     beforeEach(function (){
         module("FM.playlist.PlaylistCtrl");
@@ -12,7 +12,8 @@ describe("FM.playlist.PlaylistCtrl", function() {
      beforeEach(inject(function (_$httpBackend_) {
         $httpBackend = _$httpBackend_
 
-        $httpBackend.whenGET(/.*player\/queue/).respond(200, [{ track: { uri: "foo" } },{ track: { uri: "bar" } }]);
+        $httpBackend.whenGET(/.*player\/queue/).respond(200, [{ track: { uri: "foo", duration: 100 } },{ track: { uri: "bar", duration: 100 } }]);
+        $httpBackend.whenGET(/.*player\/queue\/meta/).respond(200, {"play_time": 20100, "genres": {"dirty south rap": 2, "pop": 12, "quiet storm": 1}, "total": 2, "users": {"fef86892-0a28-4b26-b0b3-90a1050cfffd": 2}});
         $httpBackend.whenGET(/.*users.*/).respond(200, {"family_name": "Light", "display_name": "Alex Light", "avatar_url": "http://placehold.it/400", "spotify_playlists": null, "given_name": "Alex", "id": "16369f65-6aa5-4d04-8927-a77016d0d721"});
         $httpBackend.whenGET(/.*tracks.*/).respond(200, {"album": {"id": "d7b737a9-d70b-49a9-9f42-8c204b342000", "images": [{"url": "http://placehold.it/640x629?text=Album+Art", "width": 640, "height": 629}, {"url": "http://placehold.it/300x295?text=Album+Art", "width": 300, "height": 295}, {"url": "http://placehold.it/64x63?text=Album+Art", "width": 64, "height": 63}], "name": "Boston", "uri": "spotify:album:2QLp07RO6anZHmtcKTEvSC"}, "name": "More Than a Feeling", "uri": "spotify:track:1QEEqeFIZktqIpPI4jSVSF", "play_count": 0, "artists": [{"id": "8c22640a-02ef-4ee0-90eb-87c9c9a2534f", "uri": "spotify:artist:29kkCKKGXheHuoO829FxWK", "name": "Boston"}], "duration": 285133, "id": "0739b113-ad3a-47a4-bea9-edb00ba192f5"});
         $httpBackend.whenGET(/partials\/.*/).respond(200);
@@ -34,11 +35,13 @@ describe("FM.playlist.PlaylistCtrl", function() {
 
         PlayerQueueResource = $injector.get("PlayerQueueResource");
         spyOn(PlayerQueueResource, "query").and.callThrough();
+        spyOn(PlayerQueueResource, "meta").and.callThrough();
 
         UsersResource = $injector.get("UsersResource");
         spyOn(UsersResource, "get").and.callThrough();
 
-        playlistData = { items: [{ track: { uri: "foo" } },{ track: { uri: "bar" } }]};
+        playlistData = { items: [{ track: { uri: "foo", duration: 100 } },{ track: { uri: "bar", duration: 100 } }]};
+        playlistMeta = {"play_time": 20100, "genres": {"dirty south rap": 2, "pop": 12, "quiet storm": 1}, "total": 2, "users": {"fef86892-0a28-4b26-b0b3-90a1050cfffd": 9}};
 
         $controller("PlaylistCtrl", {
             $scope: $scope,
@@ -47,7 +50,8 @@ describe("FM.playlist.PlaylistCtrl", function() {
             TracksResource: TracksResource,
             UsersResource: UsersResource,
             PlayerQueueResource: PlayerQueueResource,
-            playlistData: playlistData
+            playlistData: playlistData,
+            playlistMeta: playlistMeta
         });
     }));
 
@@ -66,12 +70,14 @@ describe("FM.playlist.PlaylistCtrl", function() {
     it("should attach data to $scope", function(){
         $httpBackend.flush();
         expect($scope.playlist).toEqual(playlistData.items);
+        expect($scope.meta).toEqual(playlistMeta);
     });
 
     it("should refresh playlist data", function(){
         $scope.refreshPlaylist();
         $httpBackend.flush();
         expect(PlayerQueueResource.query).toHaveBeenCalled();
+        expect(PlayerQueueResource.meta).toHaveBeenCalled();
     });
 
     it("should refresh playlist data on play event", function(){
@@ -94,6 +100,13 @@ describe("FM.playlist.PlaylistCtrl", function() {
         $httpBackend.flush();
     });
 
+    it("should update playlist meta on end event", function(){
+        $scope.onEnd({},{ uri: "foo" });
+        $httpBackend.flush();
+        expect($scope.meta.play_time).toEqual(20000);
+        expect($scope.meta.total).toEqual(1);
+    });
+
     it("should add item to playlist on add event", function(){
         $scope.onAdd({},{ uri: "foo", user: 123 });
         $httpBackend.flush();
@@ -101,5 +114,12 @@ describe("FM.playlist.PlaylistCtrl", function() {
         expect($notification).toHaveBeenCalledWith("Track Added", { body: "Alex Light added Boston - Boston: More Than a Feeling", icon: "http://placehold.it/640x629?text=Album+Art" });
     });
 
+    it("should update playlist meta on add event", function(){
+        var playTime = $scope.meta.play_time + 285133;
+        $scope.onAdd({},{ uri: "foo", user: 123 });
+        $httpBackend.flush();
+        expect($scope.meta.total).toEqual(3);
+        expect($scope.meta.play_time).toEqual(playTime);
+    });
 
 });